requirements.txt
flask
python-dotenv
markdown
google-generativeai
